# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sintra AI Orchestration Platform API
  description: |
    Enterprise-grade AI orchestration platform for intelligent automation and workflow management.
    
    ## Features
    - Multi-agent AI orchestration
    - Real-time chat and collaboration
    - Knowledge base management
    - Workflow automation
    - Analytics and monitoring
    
    ## Authentication
    The API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests per hour
    - AI chat endpoints: 100 requests per hour
    - Admin endpoints: 500 requests per hour
    
    ## Error Handling
    All endpoints return standardized error responses with German error messages.
    
    ## SDKs
    - JavaScript/TypeScript: `@sintra-ai/sdk`
    - Python: `sintra-ai-python`
    - Go: `github.com/sintra-ai/go-sdk`
  version: 3.0.0
  contact:
    name: Sintra AI Support
    email: support@sintra-ai.com
    url: https://sintra-ai.com/support
  license:
    name: Proprietary
    url: https://sintra-ai.com/license
  termsOfService: https://sintra-ai.com/terms
servers:
  - url: https://api.sintra-ai.com/v3
    description: Production server
  - url: https://staging-api.sintra-ai.com/v3
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /agents:
    get:
      tags:
        - Agents
      summary: List all agents
      description: Returns a list of all available AI agents
      operationId: listAgents
      parameters:
        - name: status
          in: query
          description: Filter by agent status
          schema:
            type: string
            enum: [active, inactive, maintenance]
        - name: type
          in: query
          description: Filter by agent type
          schema:
            type: string
            enum: [chat, analysis, automation, custom]
        - name: limit
          in: query
          description: Maximum number of agents to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of agents to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Agents
      summary: Create a new agent
      description: Creates a new custom AI agent
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /agents/{agentId}:
    get:
      tags:
        - Agents
      summary: Get agent by ID
      description: Returns detailed information about a specific agent
      operationId: getAgent
      parameters:
        - name: agentId
          in: path
          required: true
          description: Unique identifier of the agent
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Agents
      summary: Update agent
      description: Updates an existing agent's configuration
      operationId: updateAgent
      parameters:
        - name: agentId
          in: path
          required: true
          description: Unique identifier of the agent
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Agents
      summary: Delete agent
      description: Permanently deletes an agent
      operationId: deleteAgent
      parameters:
        - name: agentId
          in: path
          required: true
          description: Unique identifier of the agent
          schema:
            type: string
      responses:
        '204':
          description: Agent deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /agents/{agentId}/chat:
    post:
      tags:
        - Agents
      summary: Chat with agent
      description: Send a message to an AI agent and receive a response
      operationId: chatWithAgent
      parameters:
        - name: agentId
          in: path
          required: true
          description: Unique identifier of the agent
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
          text/event-stream:
            schema:
              type: string
      responses:
        '200':
          description: Chat response (streaming)
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"type": "message", "content": "Hello! How can I help you?", "agentId": "dexter"}
                  
                  data: {"type": "done"}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      tags:
        - Agents
      summary: Get agent chat history
      description: Returns the chat history for a specific agent
      operationId: getAgentChatHistory
      parameters:
        - name: agentId
          in: path
          required: true
          description: Unique identifier of the agent
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of messages to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: before
          in: query
          description: Return messages before this timestamp
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Chat history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatHistory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Knowledge Base Endpoints
...
  /analytics/agents:
    get:
      tags:
        - Analytics
      summary: Get agent analytics
      description: Returns analytics data for agents
      operationId: getAgentAnalytics
      parameters:
        - name: agentId
          in: query
          description: Specific agent ID (optional)
          schema:
            type: string
        - name: period
          in: query
          description: Time period for analytics
          schema:
            type: string
            enum: [day, week, month, quarter, year]
            default: month
      responses:
        '200':
          description: Agent analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentAnalytics'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # User Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the agent
        name:
          type: string
          description: Display name of the agent
        description:
          type: string
          description: Description of the agent's capabilities
        type:
          type: string
          enum: [chat, analysis, automation, custom]
          description: Type of the agent
        status:
          type: string
          enum: [active, inactive, maintenance]
          description: Current status of the agent
        model:
          type: string
          description: AI model used by the agent
        capabilities:
          type: array
          items:
            type: string
          description: List of agent capabilities
        configuration:
          type: object
          description: Agent-specific configuration
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        total:
          type: integer
          description: Total number of agents
        limit:
          type: integer
          description: Number of agents returned
        offset:
          type: integer
          description: Number of agents skipped
    CreateAgentRequest:
      type: object
      required:
        - name
        - type
        - model
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Display name of the agent
        description:
          type: string
          maxLength: 500
          description: Description of the agent's capabilities
        type:
          type: string
          enum: [chat, analysis, automation, custom]
          description: Type of the agent
        model:
          type: string
          description: AI model to use
        capabilities:
          type: array
          items:
            type: string
          description: List of agent capabilities
        configuration:
          type: object
          description: Agent-specific configuration
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [active, inactive, maintenance]
        capabilities:
          type: array
          items:
            type: string
        configuration:
          type: object

    # Chat Schemas
    ChatRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000
          description: The message content to send to the agent
        context:
          type: object
          description: Additional context for the conversation
        stream:
          type: boolean
          default: true
          description: Whether to stream the response
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the message
        role:
          type: string
          enum: [user, assistant, system]
          description: Role of the message sender
        content:
          type: string
          description: Message content
        timestamp:
          type: string
          format: date-time
          description: When the message was created
        metadata:
          type: object
          description: Additional message metadata
    ChatHistory:
      type: object
      properties:
        agentId:
          type: string
          description: ID of the agent
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        total:
          type: integer
          description: Total number of messages
        hasMore:
          type: boolean
          description: Whether there are more messages

    # Knowledge Schemas
...
    AgentAnalytics:
      type: object
      properties:
        agentId:
          type: string
          description: ID of the agent
        period:
          type: string
          description: Time period covered
        totalMessages:
          type: integer
          description: Total messages processed
        uniqueUsers:
          type: integer
          description: Number of unique users
        averageResponseTime:
          type: number
          description: Average response time in milliseconds
        errorRate:
          type: number
          description: Error rate (0-1)
        usageByHour:
          type: array
          items:
            type: object
            properties:
              hour:
                type: integer
                description: Hour of the day (0-23)
              count:
                type: integer
                description: Number of messages

    # User Profile Schemas
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            id:
              type: string
              description: Unique error identifier
            type:
              type: string
              description: Error type
            title:
              type: string
              description: Error title in German
            message:
              type: string
              description: Error message in German
            suggestion:
              type: string
              description: Suggested solution in German
            severity:
              type: string
              enum: [LOW, MEDIUM, HIGH, CRITICAL]
              description: Error severity level
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              description: Request identifier for tracing
            context:
              type: object
              description: Additional error context
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              id: "err_1640995200_abc123"
              type: "VALIDATION_ERROR"
              title: "Validierungsfehler"
              message: "Die bereitgestellten Daten sind ungültig."
              suggestion: "Bitte überprüfen Sie Ihre Eingaben und versuchen Sie es erneut."
              severity: "MEDIUM"
              timestamp: "2025-12-14T20:59:59.000Z"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              id: "err_1640995200_def456"
              type: "AUTHENTICATION_ERROR"
              title: "Authentifizierung fehlgeschlagen"
              message: "Sie müssen angemeldet sein, um auf diese Ressource zuzugreifen."
              suggestion: "Bitte melden Sie sich an und versuchen Sie es erneut."
              severity: "HIGH"
              timestamp: "2025-12-14T20:59:59.000Z"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              id: "err_1640995200_ghi789"
              type: "AUTHORIZATION_ERROR"
              title: "Zugriff verweigert"
              message: "Sie haben keine Berechtigung für diese Aktion."
              suggestion: "Kontaktieren Sie Ihren Administrator für weitere Unterstützung."
              severity: "HIGH"
              timestamp: "2025-12-14T20:59:59.000Z"
    NotFound:
      description: Not found - resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              id: "err_1640995200_jkl012"
              type: "NOT_FOUND_ERROR"
              title: "Ressource nicht gefunden"
              message: "Die angeforderte Ressource konnte nicht gefunden werden."
              suggestion: "Überprüfen Sie die URL oder navigieren Sie zurück zur Hauptseite."
              severity: "MEDIUM"
              timestamp: "2025-12-14T20:59:59.000Z"
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              id: "err_1640995200_mno345"
              type: "CONFLICT_ERROR"
              title: "Konflikt"
              message: "Ein Konflikt ist mit dem aktuellen Zustand der Ressource aufgetreten."
              suggestion: "Versuchen Sie es später erneut oder aktualisieren Sie die Seite."
              severity: "MEDIUM"
              timestamp: "2025-12-14T20:59:59.000Z"
    RateLimited:
      description: Rate limited - too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              id: "err_1640995200_pqr678"
              type: "RATE_LIMIT_ERROR"
              title: "Rate Limit überschritten"
              message: "Zu viele Anfragen. Bitte warten Sie einen Moment."
              suggestion: "Versuchen Sie es in ein paar Minuten erneut."
              severity: "MEDIUM"
              timestamp: "2025-12-14T20:59:59.000Z"
...